Add Car domain, repository, and API support
- Introduce `Car` and `CarId` models, including JSON encoders/decoders and schema derivation.
- Implement `CarService` trait and its live repository layer `CarRepositoryGel` with support for `getCar` and `updateCar` operations.
- Create database schema and migrations for `CarGel` with mileage field as `int64`.
- Add API endpoints (`CarEndpoints` and `CarEndpointsLive`) to handle car-related operations.
- Update `Main.scala` to include `CarEndpointsLive` in API routing stack.
- Adjust related models (`TripGel

+++ repository/src/scala/gel/invoice/models/InvoiceGel.scala
@@ -5,7 +5,7 @@ import domain.models.PersonCreate
 import domain.models.invoice.{DriverName, Invoice, InvoiceId}
 import gel.person.models.PersonCreateGel
 
-import java.lang.Short
+import java.lang.Long
 import java.time.LocalDate
 import java.util
 import java.util.UUID
@@ -21,7 +21,7 @@ class InvoiceGel @GelDeserializer() (
   gelPerson: PersonCreateGel,
   kind: String,
   // Had to put String when mileage is null, otherwise it was not working
-  mileage: String | Short,
+  mileage: String | Long,
   fileName: String,
   @GelLinkType(classOf[PersonCreateGel])
   toDriver: PersonCreateGel
@@ -34,7 +34,7 @@ class InvoiceGel @GelDeserializer() (
   def getName: String              = name
   def getPerson: PersonCreateGel   = gelPerson
   def getKind: String              = kind
-  def getMileage: String | Short   = mileage
+  def getMileage: String | Long    = mileage
   def getFileName: String          = fileName
   def getToDriver: PersonCreateGel = toDriver
 }
@@ -62,7 +62,7 @@ object InvoiceGel {
       invoiceGel.getKind,
       Option(invoiceGel.getMileage).map {
         case s: String => s.toInt
-        case l: Short  => l.toInt
+        case l: Long   => l.toInt
       },
       Option(invoiceGel.getFileName).filter(_.nonEmpty),
       Option(invoiceGel.getToDriver).map(person => DriverName(person.name))
